# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Shop Api",
        "version": "1.0.0"
    },
    "host": "shop.loc",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories with relations",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "XMLHttpRequest"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Category"
                                },
                                {
                                    "properties": {
                                        "categories_recursive": {
                                            "description": "Array of subcategories",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create category",
                "operationId": "createCategory",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "XMLHttpRequest"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "parent_id",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by id with relations",
                "description": "Get category by id",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "type": "integer",
                        "minimum": 1
                    },
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "default": "XMLHttpRequest"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Category"
                                },
                                {
                                    "properties": {
                                        "categories_recursive": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Category is not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "Need to validate",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "parent_id",
                        "in": "formData",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Category is not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/categories/{id}": {
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Remove category",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id number of category",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Category is not found"
                    }
                }
            }
        },
        "/all-categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "description": "Get categories",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Post is not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Category": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    }
}